Migration to remove version from Arxiv IDs
Most recent versions are turned into versions without an ID.
Other versions are soft deleted. At some point they could be
hard deleted, along with the source files.

A few commits prior is a change that makes the site use Arxiv IDs
without versions. It is safe to run this migration after that is
deployed because the migration just uses any existing papers
without a version if they already exist.